Minimize API surface area
