--- src/VariadicDotNotationMap/index.ts
@@ -1,4 +1,0 @@
-export {
-  VariadicDotNotationMap,
-  VariadicDotNotationKeys,
-} from "./VariadicDotNotationMap";
